+++ helpers/index.js
@@ -1,6 +1,8 @@
 module.exports = {
+    addChoicesToSelect: require('./addChoicesToSelect').addChoicesToSelect,
     batchProcess: require('./batchProcess').batchProcess,
     checkPermissions: require('./checkPermissions').checkPermissions,
     formatQuery: require('./formatQuery').formatQuery,
     wait: require('./wait').wait,
+    validateTable: require('./validateTable').validateTable,
 }