+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 __Features:__
 
 * Simplify API route structure. Condense more into the main `fetchWrapper()` function.
+* Function to `addChoicesToSelect`.
+* Function to `validateTable`.
+
 # 0.8.8 - 2022-10-06
 * Fix: remove filtering of data before batch processing. Assumed to be in record
 format.